style(lint): passe de tipagem e formatacao sobre a sessao Astra
Passe de lint e tipagem em 31 arquivos -- 18 Python, 10 do frontend, 1 HTML,
a configuracao de Jest e um setup de Jest novo. Nenhuma mudanca de
comportamento pretendida.

Auditado antes de entrar: pyright 0/0, pyrefly 0, tsc 0, eslint 0,
jest 48 suites e 381 testes, pytest 1073 aprovados e 1 pulado.

Tres regressoes do passe foram encontradas e corrigidas no mesmo commit:

- test_hybrid_router.py: o import de pydantic entrou depois da manipulacao de
  sys.path sem o noqa E402 do vizinho, e o HEAD passava limpo. Movido para o
  bloco de topo -- pydantic e pacote instalado e nao depende daquele sys.path.
  Corrigir e melhor que silenciar.
- agents/fallback.py: formatacao fora do padrao ruff, com o HEAD formatado.
- plot_benchmark.py: o passe removeu o cast que 23e9840d ja tinha,
  reintroduzindo 1 erro e 2 avisos de Pyright. A versao aqui reune as
  melhorias do passe e o cast de volta.

Vinte e cinco registros historicos ancoram nos arquivos tocados. Reconciliados
em AUDITORIA-2026-09-10-passe-de-lint-e-reconciliacao-de-ancoras.md, com id,
caminhos e parecer por item -- reconciliacao, nao dispensa.

+++ api/v1/handlers.py
@@ -944,8 +944,8 @@ async def handle_prometheus_metrics(request: web.Request) -> web.Response:
     ram_free_mb = 6400.0
     if psutil is not None:
         with contextlib.suppress(Exception):
-            cpu_load = float(psutil.cpu_percent())
-            ram_free_mb = float(psutil.virtual_memory().available / (1024 * 1024))
+            cpu_load = psutil.cpu_percent()
+            ram_free_mb = psutil.virtual_memory().available / (1024 * 1024)
 
     lines = [
         "# HELP nexus_tasks_total Total de tarefas registradas na fila por status",
